+++ tests/weird_usage.rs
@@ -7,7 +7,7 @@ use std::time::Duration;
 
 use divan::{Divan, __private::BENCH_ENTRIES};
 
-#[divan::bench(bytes_count = 0u8, chars_count = 0u16, items_count = 0u32)]
+#[divan::bench(bytes_count = 0u8, chars_count = 0u16, cycles_count = 0u32, items_count = 0u64)]
 fn zero_throughput() {}
 
 #[divan::bench(min_time = Duration::ZERO)]