+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- [`CyclesCount`] counter to display cycle throughput as Hertz.
 - Track the maximum number of bytes allocated during a benchmark.
 
 ### Fixed
@@ -313,6 +314,7 @@ Initial release. See [blog post](https://nikolaivazquez.com/blog/divan/).
 [`BytesCount::of_iter`]: https://docs.rs/divan/0.1/divan/counter/struct.BytesCount.html#method.of_iter
 [`BytesCount::of_many`]: https://docs.rs/divan/0.1/divan/counter/struct.BytesCount.html#method.of_many
 [`consts`]: https://docs.rs/divan/latest/divan/attr.bench.html#consts
+[`CyclesCount`]: https://docs.rs/divan/0.1/divan/counter/struct.CyclesCount.html
 [`Divan::max_time`]: https://docs.rs/divan/0.1/divan/struct.Divan.html#method.max_time
 
 [`Any`]: https://doc.rust-lang.org/std/any/trait.Any.html